+++ Runtime.sln
@@ -91,8 +91,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Metrics", "Specifications\M
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Events.Streams", "Source\Events.Streams\Events.Streams.csproj", "{36EC38EF-BCF5-401D-B1E1-E4AF1924F91A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Events.Streams.Processing", "Source\Events.Streams.Processing\Events.Streams.Processing.csproj", "{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -610,18 +608,6 @@ Global
 		{36EC38EF-BCF5-401D-B1E1-E4AF1924F91A}.Release|x64.Build.0 = Release|Any CPU
 		{36EC38EF-BCF5-401D-B1E1-E4AF1924F91A}.Release|x86.ActiveCfg = Release|Any CPU
 		{36EC38EF-BCF5-401D-B1E1-E4AF1924F91A}.Release|x86.Build.0 = Release|Any CPU
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}.Debug|x64.Build.0 = Debug|Any CPU
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}.Debug|x86.Build.0 = Debug|Any CPU
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}.Release|x64.ActiveCfg = Release|Any CPU
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}.Release|x64.Build.0 = Release|Any CPU
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}.Release|x86.ActiveCfg = Release|Any CPU
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{A8DB2A0A-B52B-4B70-A201-29C40E6ABDCE} = {6DECF56C-D430-4E51-AB9A-59FD49309EEE}
@@ -666,6 +652,5 @@ Global
 		{2F737745-EE5C-4E6E-AAFD-70852A1465E6} = {832EF7FD-6E0C-488D-8CA1-3D5AB0F7CBA2}
 		{C77CC3BC-E19C-4515-BC50-F09F785190A2} = {6DECF56C-D430-4E51-AB9A-59FD49309EEE}
 		{36EC38EF-BCF5-401D-B1E1-E4AF1924F91A} = {832EF7FD-6E0C-488D-8CA1-3D5AB0F7CBA2}
-		{B0BFFCEA-A451-4339-8CEF-3CE2CDFE5B10} = {832EF7FD-6E0C-488D-8CA1-3D5AB0F7CBA2}
 	EndGlobalSection
 EndGlobal
