--- Source/Events.Streams.Processing/Events.Streams.Processing.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <Import Project="../../default.props"/>
-
-  <PropertyGroup>
-    <AssemblyName>Dolittle.Runtime.Events.Streams.Processing</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Events\Events.csproj" />
-    <ProjectReference Include="..\Events.Streams\Events.Streams.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Dolittle.Concepts" Version="4.*" />
-    <PackageReference Include="System.Reactive" Version="4.3.2" />
-  </ItemGroup>
-</Project>
