+++ Source/Events.Streams/Processing/EventProcessors.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Dolittle.Runtime.Events.Streams
+namespace Dolittle.Runtime.Events.Streams.Processing
 {
     /// <summary>
     /// Represents an implementation of <see cref="IEventProcessors"/>.
