+++ Source/Events.Streams/Processing/IEventProcessors.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Dolittle.Runtime.Events.Streams
+namespace Dolittle.Runtime.Events.Streams.Processing
 {
     /// <summary>
     /// Defines a system that can process events.
